% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.r
\name{A}
\alias{A}
\title{A}
\usage{
A(x)
}
\description{
This function is a shorthand for \code{attr()}.
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
\code{\link[base]{attr}}

Other poorly documented: \code{\link{applySum}},
  \code{\link{assignStation}}, \code{\link{colour.scale}},
  \code{\link{compileODF}},
  \code{\link{compute.oxygen.saturation}},
  \code{\link{compute.sums}}, \code{\link{concave.hull}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}, \code{\link{cor.prob}},
  \code{\link{correlationMatrix}}
}

