% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyMean.r
\name{applyMean}
\alias{applyMean}
\title{applyMean}
\usage{
applyMean(f, method = "fast", newnames = NULL, ...)
}
\description{
apply a mean or weighted mean to data by factor
and return a formatted data.frame
must contain index, variable, weight in this sequence 
can take additional parmaters ...
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other poorly documented, unclaimed: \code{\link{applySum}}
}

