% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applySum.r
\name{applySum}
\alias{applySum}
\title{applySum}
\usage{
applySum(f, method = "fast", newnames = NULL)
}
\description{
apply a sum and return a formated data frame
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
 @export
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{assignStation}}, \code{\link{colour.scale}},
  \code{\link{compileODF}},
  \code{\link{compute.oxygen.saturation}},
  \code{\link{compute.sums}}, \code{\link{concave.hull}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}, \code{\link{cor.prob}},
  \code{\link{correlationMatrix}}
}

