% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio.blue.r
\name{bio.blue}
\alias{bio.blue}
\title{bio.blue}
\usage{
bio.blue(n, alpha = 1)
}
\arguments{
\item{\code{n}}{= the number of colors desired}

\item{\code{alpha}}{= the transparency desired (default is fully opaque/1)}
}
\value{
character vector
}
\description{
This function generates a vector of shades of blue (from white to blue) of the length selected by value of \code{n}.
}
\examples{
bio.blue(5)
[1] "#FFFFFFFF" "#AAAAFFFF" "#5555FFFF" "#0000FFFF"
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
