% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue.red.r
\name{blue.red}
\alias{blue.red}
\title{blue.red}
\usage{
blue.red(n)
}
\arguments{
\item{\code{n}}{= the number of colors desired}
}
\value{
character vector
}
\description{
This function generates a vector of shades from blue to red of a length selected by value of \code{n}.
}
\note{
As of 2016-06-14, this function generates an error: unused argument (gamma = gamma).  It appears to be attempting to generate a hsv value by adding a gamma value at the end of a hex value.  Removing \code{, gamma=gamma} generates the expected ramp.
}
\examples{
blue.red(5)
[1] "#0000FF" "#FFFFFF" "#FFFFFF" "#FF0000"
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
