% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufferCircle.r
\name{bufferCircle}
\alias{bufferCircle}
\title{bufferCircle}
\usage{
bufferCircle(lonlat, radius, npoints = 100)
}
\arguments{
\item{\code{lonlat}}{= the position to buffer}

\item{\code{radius}}{= the radius in kilometers to buffer}

\item{\code{npoints}}{= number of discrete points to create (default is 100)}
}
\value{
data.frame of 101 indivual points of $lat and $lon
}
\description{
This function buffers a point by a specified radius \code{n}.
}
\examples{
bufferCircle(c(-64.15,44.75),100)
}
\author{
Adam Cook, \email{Adam.Cook@dfo-mpo.gc.ca}
}
