% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufferRange.r
\name{bufferRange}
\alias{bufferRange}
\title{bufferRange}
\usage{
bufferRange(x, buffer.percent = 2)
}
\arguments{
\item{\code{buffer.percent}}{= buffer amount (default = 2)}
}
\value{
2 values, each buffered to extend the range by the selected percentage
}
\description{
This function adds a bit of negative a positive buffer to a range.
}
\examples{
x=-1000:10   
bufferRange(x,2)
[1] -1020.0    10.2
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
