% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by2.r
\name{by2}
\alias{by2}
\title{@title by2
 @description  'slower but more flexible'
 @family poorly documented
 @author  unknown, \email{<unknown>@dfo-mpo.gc.ca}
 @export}
\usage{
by2(x, indices, func, var, newvars, ...)
}
\description{
@title by2
 @description  'slower but more flexible'
 @family poorly documented
 @author  unknown, \email{<unknown>@dfo-mpo.gc.ca}
 @export
}

