% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap.r
\name{cap}
\alias{cap}
\title{cap}
\usage{
cap(s, strict = FALSE)
}
\arguments{
\item{\code{s}}{= the string to handle}

\item{\code{strict}}{= if \code{TRUE}, removes all other uppercase letters; if \code{FALSE} (the default), ignore subsequent capitals}
}
\value{
character vector
}
\description{
This function ensures that the passed string gets a capitalized letter
}
\examples{
cap('this is a TEST', F)
[1] "This is a TEST"
cap('this is a TEST', T)
[1] "This is a test"
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
