% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capwords.r
\name{capwords}
\alias{capwords}
\title{capwords}
\usage{
capwords(s, strict = FALSE, sentence = T)
}
\arguments{
\item{\code{s}}{= the string to handle}

\item{\code{strict}}{= if \code{TRUE}, forces capitalization of all words (as long as \code{sentence = F})}

\item{\code{sentence}}{= if \code{TRUE}, only the first word is capitalized}
}
\value{
character vector
}
\description{
This function capitalizes words
}
\note{
taken from help of 'toupper'
}
\examples{
capwords('this is a test', F, T)
[1] "This is a test"
capwords('this is a test', T, T)
[1] "This is a test"
capwords('this is a test', T, F)
[1] "This Is A Test"
capwords('this is a test', F, F)
[1] "This Is A Test"
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
