% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron2datestamp.r
\name{chron2datestamp}
\alias{chron2datestamp}
\title{chron2datestamp}
\usage{
chron2datestamp(x, res = 12)
}
\arguments{
\item{\code{x}}{= the date}

\item{\code{res}}{= the time will be calculated as a fraction of this value (defaults to \code{12})}
}
\value{
character vector
}
\description{
This function converts dates from \code{chron::} format and append the time as a fraction, for use as a timestamp
}
\examples{
require(chron)
t1<- chron(dates =  dates(c("02/27/92")), times =  "08:00:00")
chron2datestamp(t1,100)
[1] "02/27/92.33"

t2<- chron(dates =  dates(c("02/27/92")), times =  "12:00:00")
chron2datestamp(t2,10)
[1] "02/27/92.5"
}
\seealso{
Other date functions: \code{\link{convert.datecodes}},
  \code{\link{date_picker}}, \code{\link{dayofyear}},
  \code{\link{get.day}}, \code{\link{get_day}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
