% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.memory.r
\name{clear.memory}
\alias{clear.memory}
\title{clear.memory}
\usage{
clear.memory()
}
\description{
This function clears all data sets and functions defined in the current environment
}
\note{
taken from help of 'toupper'
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
