% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipB.r
\name{clipB}
\alias{clipB}
\title{clipB}
\usage{
clipB(x, sep = "\\t", row.names = FALSE, col.names = TRUE)
}
\description{
Writes table into clipboard in unix environment. Requires xclip be installed (pacman xclip).
}
\author{
Stackoverflow
}
