% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.code.r
\name{color.code}
\alias{color.code}
\title{color.code}
\usage{
color.code(type = "seis", n)
}
\arguments{
\item{\code{type}}{= the colour palette to use.  Valid options include: yellow.red; blue.black; red.white.blue; seis; red.yellow.blue; nathalie; blue.yellow.red; blue.yellow.blue; colourblind1}

\item{\code{n}}{= the number of colours to return}
}
\value{
vector of \code{n} length of hex color codes
}
\description{
This functions return a vector of colour codes using a number of pre-defined palettes ('types')
}
\examples{
color.code(type='colourblind1',n=2)
[1] "#FFFFD8" "#071D58"
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
