% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.names.r
\name{colour.names}
\alias{colour.names}
\title{colour.names}
\usage{
colour.names(colour.order = "r")
}
\arguments{
\item{\code{colour.order}}{= the sort order.  Valid values are \code{r}, \code{g} and \code{b} (\code{r} is the default)}
}
\description{
This function returns all of the named colours used by R with their associated R, G and B values
}
\examples{
color.code(type='colourblind1',n=2)
[1] "#FFFFD8" "#071D58"
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
