% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.scale.r
\name{colour.scale}
\alias{colour.scale}
\title{colour.scale}
\usage{
colour.scale(type = "seis", nlevels = 10, x = NULL, transparency = 0.9)
}
\arguments{
\item{\code{type}}{= default is \code{seis}}

\item{\code{nlevels}}{= the number of unique colours to generate (default is 10)}

\item{\code{x}}{= a vector}

\item{\code{transparency}}{= the transparency of the resultant colours (default is 0.9)}
}
\value{
list of 2: \code{$y} is a vector of \code{x-1} length, which can used to assign the colours of \code{$cols} to the original vector \code{x}; \code{$cols} is a vector of KML colours of length \code{nlevels}
}
\description{
This functions return a vector of length \code{x} KML (Google Earth) colour codes using a pre-defined palettes ('seis').
}
\note{
not totally sure if this function works as intended.  It seems as though \code{nlevels} should return the number of colours specified - not one less.  Jae or Adam might have more information about this.
}
\examples{
colour.scale(nlevels=3,x=seq(1:6))
$y
[1] 1 1 2 2 3 3

$cols
code
1 e68B0000
2 e600008B
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applyMean}}, \code{\link{applySum}},
  \code{\link{break.list}}, \code{\link{by2}},
  \code{\link{compute.sums}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}, \code{\link{dim_list}},
  \code{\link{duplicates.toremove}}, \code{\link{na.zero}},
  \code{\link{nearest.sorted.angles}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
