% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileODF.r
\name{compileODF}
\alias{compileODF}
\title{compileODF}
\usage{
compileODF(path = loc)
}
\arguments{
\item{\code{path}}{= location of an ODF file}
}
\value{
\code{result}
}
\description{
No description. Requires the \code{parse.odf.file()} function from this package.
}
\note{
MMM - 2016-06-17 - Unable to get this to work.
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applyMean}}, \code{\link{applySum}},
  \code{\link{break.list}}, \code{\link{by2}},
  \code{\link{colour.scale}},
  \code{\link{compute.oxygen.saturation}},
  \code{\link{compute.sums}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}
}

