% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.oxygen.saturation.r
\name{compute.oxygen.saturation}
\alias{compute.oxygen.saturation}
\title{compute.oxygen.saturation}
\usage{
compute.oxygen.saturation(t.C, sal.ppt, oxy.ml.l)
}
\arguments{
\item{\code{t.C}}{= temperature in Celsius}

\item{\code{sal.ppt}}{= salinity in ppt}

\item{\code{oxy.ml.l}}{= oxygen saturation in ml/L}
}
\value{
\code{DO} = saturation percentage
}
\description{
This function determines the percentage of Dissolved oxygen in solution at a given temperature and salinity given a known saturation amount. formula attributed to Weiss DSR 17, 721-735, 1970
}
\examples{
compute.oxygen.saturation(0, 25, 8.2)
[1] 95.19575
}
\seealso{
Other physical and chemical properties: \code{\link{decibar2depth}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
