% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.sums.r
\name{compute.sums}
\alias{compute.sums}
\title{compute.sums}
\usage{
compute.sums(x, var, index)
}
\arguments{
\item{\code{x}}{=the input matrix}

\item{\code{var}}{=}

\item{\code{index}}{}
}
\value{
\code{res}
}
\description{
This function takes a matrix, and for a given column and index, returns a matrix of different unique values for each.
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applyMean}}, \code{\link{applySum}},
  \code{\link{break.list}}, \code{\link{by2}},
  \code{\link{colour.scale}}, \code{\link{compileODF}},
  \code{\link{compute.oxygen.saturation}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}
}

