% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concave.hull.r
\name{concave.hull}
\alias{concave.hull}
\title{concave.hull}
\usage{
concave.hull(xy, k = 5, ub = NULL, dname = NULL, fname = NULL,
  random.start = F)
}
\arguments{
\item{\code{xy}}{}

\item{\code{k}}{defaults to 5}

\item{\code{ub}}{defaults to NULL}

\item{\code{dname}}{defaults to NULL}

\item{\code{fname}}{defaults to NULL}

\item{\code{random.start}}{defaults to FALSE}
}
\description{
cannot accurately describe
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other abysmally documented: \code{\link{cor.prob}},
  \code{\link{create.histograms.yearly}},
  \code{\link{errorbars}}
}

