% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_ptran.r
\name{connect_ptran}
\alias{connect_ptran}
\title{connect_ptran}
\usage{
connect_ptran(user = "x", password = "x", i = 0)
}
\arguments{
\item{\code{user}}{}

\item{\code{password}}{}

\item{\code{i}}{}
}
\value{
\code{channel}
}
\description{
This function is supposed to be a user-friendly means for adding a connection to ptran.  The user is prompted for the credentials 5 times before it
}
\seealso{
Other GUI functions: \code{\link{connect.ptran}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
