% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.degdec2degmin.r
\name{convert.degdec2degmin}
\alias{convert.degdec2degmin}
\title{convert.degdec2degmin}
\usage{
convert.degdec2degmin(x)
}
\arguments{
\item{\code{x}}{= dataframe to be handled, with lat and lon colnames}
}
\description{
This function converts decimals to degree and minute notation
}
\examples{
test=as.data.frame(cbind(42.5,55.255))
colnames(test)=c("lat","lon")
convert.degdec2degmin(test)
lat    lon
1 42.3 55.153
(55.153 is 55deg 15.3min)
}
\seealso{
Other coordinate converters: \code{\link{convert.dd.dddd}},
  \code{\link{convert.degmin2degdec}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
