% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.degmin2degdec.r
\name{convert.degmin2degdec}
\alias{convert.degmin2degdec}
\title{convert.degmin2degdec}
\usage{
convert.degmin2degdec(x, y = NULL, vnames = c("lon", "lat"))
}
\arguments{
\item{\code{x}}{= dataframe to be handled, with lat and lon colnames}
}
\description{
This function converts coordinates in dd.mmmm to decimal degrees
}
\examples{
test=as.data.frame(cbind(-52.3050,55.20))
colnames(test)=c("lat","lon")
convert.degmin2degdec(test)
lat      lon
1 -52.50833 55.33333
}
\seealso{
Other coordinate converters: \code{\link{convert.dd.dddd}},
  \code{\link{convert.degdec2degmin}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
