% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.graphics.format.r
\name{convert.graphics.format}
\alias{convert.graphics.format}
\title{convert.graphics.format}
\usage{
convert.graphics.format(file0, file1, options = "")
}
\arguments{
\item{\code{file0}}{}

\item{\code{file1}}{}

\item{\code{options}}{}
}
\value{
\code{}
}
\description{
this uses imagemagicks functionality
}
\note{
does not work in Windows
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\references{
\url{http://www.imagemagick.org/script/convert.php}
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applySum}}, \code{\link{assignStation}},
  \code{\link{colour.scale}}, \code{\link{compileODF}},
  \code{\link{compute.oxygen.saturation}},
  \code{\link{compute.sums}}, \code{\link{concave.hull}},
  \code{\link{convert2factor}}, \code{\link{cor.prob}},
  \code{\link{correlationMatrix}}
}

