% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2factor.r
\name{convert2factor}
\alias{convert2factor}
\title{convert2factor}
\usage{
convert2factor(x, vars)
}
\arguments{
\item{\code{x}}{}

\item{\code{vars}}{}
}
\value{
\code{}
}
\description{
converts objects to factors
}
\note{
does not work in Windows
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{assignStation}}, \code{\link{colour.scale}},
  \code{\link{compileODF}},
  \code{\link{compute.oxygen.saturation}},
  \code{\link{compute.sums}}, \code{\link{concave.hull}},
  \code{\link{convert.graphics.format}},
  \code{\link{cor.prob}}, \code{\link{correlationMatrix}}
}

