% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.data.structure.r
\name{copy.data.structure}
\alias{copy.data.structure}
\title{copy.data.structure}
\usage{
copy.data.structure(x)
}
\arguments{
\item{\code{x}}{the source data.frame}
}
\value{
\code{y} the empty data.frame
}
\description{
takes an exiting data frame, and returns the structure only
}
\note{
MMM - 2016-06-16 - a note in the script says 'there must be an easier way of doing this ?' and there is:
y=x[0,] - this gets the zero'th row
}
\examples{
n = c(2, 3, 5) 
s = c("aa", "bb", "cc") 
b = c(TRUE, FALSE, TRUE) 
df = data.frame(n, s, b) 
str(copy.data.structure(df))
'data.frame':	0 obs. of  3 variables:
$ n: num 
$ s: chr 
$ b: logi 
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
