% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.prob.r
\name{cor.prob}
\alias{cor.prob}
\title{cor.prob}
\usage{
cor.prob(X, dfr = nrow(X) - 2)
}
\arguments{
\item{\code{X}}{}

\item{\code{dfr}}{degrees of freedom}
}
\value{
\code{}
}
\description{
couldn't accurately describe
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{assignStation}}, \code{\link{colour.scale}},
  \code{\link{compileODF}},
  \code{\link{compute.oxygen.saturation}},
  \code{\link{compute.sums}}, \code{\link{concave.hull}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}},
  \code{\link{correlationMatrix}}
}

