% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.histograms.yearly.r
\name{create.histograms.yearly}
\alias{create.histograms.yearly}
\title{create.histograms.yearly}
\usage{
create.histograms.yearly(X, probs = c(0, 0.995))
}
\arguments{
\item{\code{X}}{}

\item{\code{probs}}{}
}
\value{
\code{nn}
}
\description{
couldn't accurately describe
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other abysmally documented: \code{\link{concave.hull}},
  \code{\link{cor.prob}}, \code{\link{errorbars}}
}

