% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_matrix.r
\name{create_random_matrix}
\alias{create_random_matrix}
\title{create_random_matrix}
\usage{
create_random_matrix(the.nrow = 10, the.ncol = 2, the.range = 1:50)
}
\arguments{
\item{\code{the.nrow}}{= desired number of rows (default is 5)}

\item{\code{the.ncol}}{= desired number of columns (default is 5)}

\item{\code{the.range}}{= the range of values to use (default is 1:100)}
}
\value{
matrix of integers with the specified dimensions
}
\description{
This function calculates depth in meters from pressure in decibars using Saunders and Fofonoff's method
}
\examples{
create.random.matrix(the.nrow=10, the.ncol=20, the.range=0:9)
}
\seealso{
Other random data: \code{\link{createTS}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
