% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTS.r
\name{createTS}
\alias{createTS}
\title{taken from http://stats.stackexchange.com/questions/29239/creating-auto-correlated-random-values-in-r'}
\usage{
createTS(N = 100, seed = NULL, nonstationary = T, all.positive = T)
}
\description{
taken from http://stats.stackexchange.com/questions/29239/creating-auto-correlated-random-values-in-r'
}

