% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutOutLand.r
\name{cutOutLand}
\alias{cutOutLand}
\title{cutOutLand}
\usage{
cutOutLand(Polys, coast, mapRes = "UR", keepUniquePolys = T)
}
\arguments{
\item{\code{Polys}}{}

\item{\code{coast}}{(Defaults to \code{null})}

\item{\code{mapRes}}{(Defaults to \code{UR})}

\item{\code{keepUniquePolys}}{= if you want all values positive set all.positive=T to rescale (Defaults to T)}
}
\value{
\code{x}
}
\description{
PBS mapping utility to remove the land from management area polygons, include coastline data or defaults to "bio.data/lobster/data/maps/gshhs/shorelineUR.csv"
}
\examples{
example.function() returns \\code{4}
example.function(6) returns \\code{36}.
}
\seealso{
Other overly specific: \code{\link{dataOverlap}},
  \code{\link{default.lattice.options}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
