% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cX.r
\name{cX}
\alias{cX}
\title{cX}
\usage{
cX(X, row.names = F)
}
\arguments{
\item{\code{X}}{}

\item{\code{row.names}}{should row names be copied too ? (Defaults to \code{F})}
}
\value{
\code{}
}
\description{
write table to paste into excel
}
\examples{
cX(A)
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
