% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataOverlap.r
\name{dataOverlap}
\alias{dataOverlap}
\title{dataOverlap}
\usage{
dataOverlap(data1, data2, distance.Theshold = 0.1, dnames = c("EID", "X",
  "Y", "date", "salinity", "temperature", "depth", "source"))
}
\arguments{
\item{\code{data1}}{}

\item{\code{data2}}{}

\item{\code{distance.Theshold}}{}
}
\value{
\code{Data}
}
\description{
compare 2 datasets and returns values that overlap
}
\note{
the return object should probably be Data1, not Data
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other overly specific: \code{\link{assignStation}},
  \code{\link{cutOutLand}}
}

