% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decibar2depth.r
\name{decibar2depth}
\alias{decibar2depth}
\title{decibar2depth}
\usage{
decibar2depth(P, lat, Del = 0, method = "default")
}
\arguments{
\item{\code{P}}{= pressure, in decibars}

\item{\code{lat}}{= degrees latitude}

\item{\code{Del}}{= geopotential anomaly (defaults to 0)}

\item{\code{method}}{= valid values include"default", "seabird", and "unesco", although they all appear to be treated the same}
}
\value{
\code{DEPTH} = depth in M
}
\description{
This function calculates depth in meters from pressure in decibars using Saunders and Fofonoff's method
}
\note{
/ UNITS:
//   PRESSURE     P      DECIBARS
//   LATITUDE     LAT    DEGREES
//   DEPTH        DEPTH  METERS
//   DTN. HEIGHT  DEL    DYN. METERS (geopotenial anomaly in J/kg:: assume 0) 
// CHECKVALUE:
//   1.) DEPTH = 9712.653 M for P=10000 DECIBARS, LAT=30 DEG, DEL=0
//   ABOVE FOR STANDARD OCEAN: T=0 DEG CELSIUS; S=35 (PSS-78)
// ----------------------------------------------------------
// Original fortran code is found in:
//   UNESCO technical papers in marine science 44 (1983) -
//   'Algorithms for computation of fundamental properties of seawater'
}
\examples{
decibar2depth(10000,30,0)
[1] 9712.653
}
\references{
\url{http://www.code10.info/index.php?option=com_content&view=article&id=67:calculating-the-depth-from-pressure&catid=54:cat_coding_algorithms_seawater&Itemid=79}, \url{http://www.code10.info/index.php?option=com_content&view=article&id=67:calculating-the-depth-from-pressure&catid=54:cat_coding_algorithms_seawater&Itemid=79}
}
\seealso{
Other physical and chemical properties: \code{\link{compute.oxygen.saturation}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
