% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.lattice.options.r
\docType{data}
\name{default.lattice.options}
\alias{default.lattice.options}
\title{default.lattice.options}
\format{An object of class \code{list} of length 44.}
\usage{
default.lattice.options
}
\description{
couldn't accurately describe
}
\note{
Seems like using this function might just add more complexity rather than calling it directly from lattice?
}
\seealso{
Other overly specific: \code{\link{cutOutLand}},
  \code{\link{dataOverlap}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\keyword{datasets}
