% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devianceGLM.r
\name{devianceGLM}
\alias{devianceGLM}
\title{devianceGLM}
\usage{
devianceGLM(x, data, test = "F", digits = 2)
}
\arguments{
\item{\code{x}}{the glm model object}

\item{\code{data}}{}

\item{\code{test}}{defaults to FALSE}

\item{\code{digits}}{defaults to 2}
}
\value{
\code{result}
}
\description{
couldn't accurately describe
}
\seealso{
Other statistical: \code{\link{correlationMatrix}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
