% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.list.r
\name{dim.list}
\alias{dim.list}
\title{dim.list}
\usage{
\method{dim}{list}(list1)
}
\arguments{
\item{\code{list1}}{}
}
\value{
\code{ab}
}
\description{
takes a list and returns a simplified list of its dimensions
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applyMean}}, \code{\link{applySum}},
  \code{\link{break.list}}, \code{\link{by2}},
  \code{\link{colour.scale}}, \code{\link{compute.sums}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}},
  \code{\link{duplicates.toremove}}, \code{\link{na.zero}}
}

