% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.plt.r
\name{dist.plt}
\alias{dist.plt}
\title{dist.plt}
\usage{
dist.plt(dist = "beta", par1, par2, x, xl = c(0, 1), col = 1, add = F,
  alpha = 0.05, plot.lines = F, title = "")
}
\arguments{
\item{\code{dist}}{}

\item{\code{par1}}{}

\item{\code{par2}}{}

\item{\code{x}}{}

\item{\code{xl}}{}

\item{\code{col}}{}
}
\value{
\code{add}

\code{alpha}

\code{plot.lines}
}
\description{
plots distributions in R, useful for setting up priors in Bayesian analyses
}
\seealso{
Other plotting: \code{\link{dstamp}},
  \code{\link{ellipse.2d}}, \code{\link{get.asp}},
  \code{\link{legend.box}}, \code{\link{legend.bubble}},
  \code{\link{pie.add}}, \code{\link{pie.draw}},
  \code{\link{pie.legend}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
