% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstamp.r
\name{dstamp}
\alias{dstamp}
\title{dstamp}
\usage{
dstamp(txt, pwd = FALSE, time. = TRUE)
}
\arguments{
\item{\code{txt}}{}

\item{\code{pwd}}{}

\item{\code{time.}}{}
}
\description{
function to add a datestamp to the bottom right corner of a figure
}
\seealso{
Other plotting: \code{\link{dist.plt}},
  \code{\link{ellipse.2d}}, \code{\link{get.asp}},
  \code{\link{legend.box}}, \code{\link{legend.bubble}},
  \code{\link{pie.add}}, \code{\link{pie.draw}},
  \code{\link{pie.legend}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
