% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.toremove.r
\name{duplicates.toremove}
\alias{duplicates.toremove}
\title{duplicates.toremove}
\usage{
duplicates.toremove(x)
}
\arguments{
\item{\code{m}}{The matrix to examine}
}
\value{
the indices of values that have duplicates
}
\description{
This function looks for rows that contain identical values as other rows, irrespective of position
}
\examples{
duplicates.toremove(c(1,2,3,4,5))
integer(0)
duplicates.toremove(c(1,1,3,4,5))
[1] 1 2

m
[,1] [,2] [,3]
[1,]    2    3    3
[2,]    2    1    2
[3,]    3    3    2
[4,]    2    1    1
[5,]    2    1    2
#duplicates.toremove(m)
[1]  1  2  4  5 12 13 15
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applyMean}}, \code{\link{applySum}},
  \code{\link{break.list}}, \code{\link{by2}},
  \code{\link{colour.scale}}, \code{\link{compute.sums}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}, \code{\link{dim_list}},
  \code{\link{na.zero}},
  \code{\link{nearest.sorted.angles}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
