% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.2d.r
\name{ellipse.2d}
\alias{ellipse.2d}
\title{ellipse.2d}
\usage{
ellipse.2d(x, y, pv = 0.05, sc = 50)
}
\arguments{
\item{\code{x}}{}

\item{\code{y}}{}

\item{\code{pv}}{the density of the region to be outlined - defaults to 95 percent (i.e. 0.05)}

\item{\code{sc}}{a scaling factor? defaults to 50}
}
\description{
This function calculates a 2D kernel density then plot highest density regions with contour().
}
\references{
\url{https://stat.ethz.ch/pipermail/r-help/2000-March/005973.html}
}
\seealso{
Other plotting: \code{\link{dist.plt}},
  \code{\link{dstamp}}, \code{\link{get.asp}},
  \code{\link{legend.box}}, \code{\link{legend.bubble}},
  \code{\link{pie.add}}, \code{\link{pie.draw}},
  \code{\link{pie.legend}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
