% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbars.r
\name{errorbars}
\alias{errorbars}
\title{errorbars}
\usage{
errorbars(b, v, br = "Sturges", nfilter = 0, lowess = 0.4, trim = 0,
  xlab, ylab, db, ...)
}
\arguments{
\item{\code{b}}{}

\item{\code{v}}{}

\item{\code{br}}{br is the number of bins to be created from the x-values}

\item{\code{nfilter}}{defaults to 0}

\item{\code{lowess}}{defaults to 0.4}

\item{\code{trim}}{defaults to 0}
}
\description{
create plots of mean y after binning x values, adding 1 SE bars
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other abysmally documented: \code{\link{concave.hull}},
  \code{\link{cor.prob}},
  \code{\link{create.histograms.yearly}},
  \code{\link{estimParamsFromCI}}
}

