% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam2matrix.r
\name{gam2matrix}
\alias{gam2matrix}
\title{gam2matrix}
\usage{
gam2matrix(oo, i = 1, n2 = 100)
}
\description{
unknown
}
\seealso{
Other abysmally documented: \code{\link{CurlyBraces}},
  \code{\link{FB2C}}, \code{\link{Pr}},
  \code{\link{Rsquared_glm}}, \code{\link{Rsquared_lm}},
  \code{\link{Rsquared}}, \code{\link{SQL_in_noquotes}},
  \code{\link{SQL_in}}, \code{\link{SU}},
  \code{\link{Tabulate}}, \code{\link{U2F}},
  \code{\link{U}}, \code{\link{cor.prob}},
  \code{\link{create.histograms.yearly}},
  \code{\link{errorbars}}, \code{\link{estimParamsFromCI}},
  \code{\link{expandSet}},
  \code{\link{explicit.to.implicit.2d.line}},
  \code{\link{export.to.clipboard}},
  \code{\link{factor2character}},
  \code{\link{factor2number}}, \code{\link{fillNaDf2}},
  \code{\link{filter.dayno}}, \code{\link{filter.season}},
  \code{\link{findCurvature}}, \code{\link{findMax}},
  \code{\link{findMoments}}, \code{\link{fix.permissions}},
  \code{\link{fm.m}}, \code{\link{geomean}},
  \code{\link{get.cpu.usage}}, \code{\link{getCols}},
  \code{\link{getPackage}}, \code{\link{goff}},
  \code{\link{graphics.convert}}, \code{\link{green.red}},
  \code{\link{gxa}}, \code{\link{gx}},
  \code{\link{hist_bivariate}}, \code{\link{idHoles}},
  \code{\link{interpolate.xy.robust}},
  \code{\link{is_leap_year}}, \code{\link{jackknifeCPUE}},
  \code{\link{jae.hist}}, \code{\link{keepOneDuplicated}},
  \code{\link{kml.background.make}},
  \code{\link{kml.background}},
  \code{\link{kml.coordinate}}, \code{\link{kml.document}},
  \code{\link{kml.end}}, \code{\link{kml.folder.end}},
  \code{\link{kml.folder.start}}, \code{\link{kml.folder}},
  \code{\link{kml.line.make}}, \code{\link{kml.line}},
  \code{\link{kml.placemark.make}},
  \code{\link{kml.placemark}},
  \code{\link{kml.polygon.make}},
  \code{\link{kml.polygon}},
  \code{\link{kml.screen.overlay}},
  \code{\link{kml.snowcrab.background}},
  \code{\link{kml.start}}, \code{\link{kml.style}},
  \code{\link{lettersToNumbers}},
  \code{\link{line.intersection.2d}},
  \code{\link{list.names.to.columns}},
  \code{\link{loadPackages}}, \code{\link{ls.objects}},
  \code{\link{lsmem}}, \code{\link{lsos}},
  \code{\link{make.histograms}}, \code{\link{make.hist}},
  \code{\link{make.list}},
  \code{\link{make.random.string}},
  \code{\link{makeMapBasic}}, \code{\link{makeMap}},
  \code{\link{makePBS}}, \code{\link{matrix.multiply}},
  \code{\link{matrixBubbles}}, \code{\link{mave}},
  \code{\link{mergePolys}},
  \code{\link{mod.graphics.basic}},
  \code{\link{mod.graphics}}, \code{\link{modes}},
  \code{\link{mult.windows}}, \code{\link{myrepo}},
  \code{\link{nlsBooter}}, \code{\link{odds}},
  \code{\link{offset.determine}},
  \code{\link{package.install}},
  \code{\link{package.list}}, \code{\link{parallel.run}},
  \code{\link{parse.odf.file}},
  \code{\link{parse.odf.parameter.header}},
  \code{\link{pause.cpu}}, \code{\link{pause}},
  \code{\link{percentDifference}},
  \code{\link{plot3AgeComparisons}},
  \code{\link{plot_histograms_temporal}},
  \code{\link{plot_nothing}},
  \code{\link{plot_timeseries}},
  \code{\link{point2PolygonDist}},
  \code{\link{pressure2Depth}},
  \code{\link{print_command}},
  \code{\link{pvalue.rsquare}},
  \code{\link{quantileBreak2NA}},
  \code{\link{quantile_estimate}},
  \code{\link{quantile_to_normal}}, \code{\link{rX}},
  \code{\link{ra.jit}}, \code{\link{rbg2bgr}},
  \code{\link{recode.time.block}},
  \code{\link{recode.time}}, \code{\link{recode}},
  \code{\link{remove.files}},
  \code{\link{removeDuplicateswithNA}},
  \code{\link{rename.df}}, \code{\link{rm.from.list}},
  \code{\link{row_to_html_table}},
  \code{\link{set_global}},
  \code{\link{setup.lattice.options}},
  \code{\link{simulateMixtures}}, \code{\link{solver}},
  \code{\link{sort_data_frame}},
  \code{\link{sort_list_amc}},
  \code{\link{spatpoly_to_spatpolydf}},
  \code{\link{sql.find.bio}}, \code{\link{squeezeBlanks}},
  \code{\link{stats.by.factors}},
  \code{\link{string2chron}}, \code{\link{stringsplit}},
  \code{\link{summary_deviance}}, \code{\link{table.view}},
  \code{\link{tapply2}}, \code{\link{test.intersection}},
  \code{\link{toNums}}, \code{\link{topo.colors2}},
  \code{\link{trimWhiteSpace}},
  \code{\link{truncate_distribution}},
  \code{\link{truncate_vector}},
  \code{\link{ts.plotandsave}},
  \code{\link{variable.list.expand}},
  \code{\link{variable.recode}}, \code{\link{which.noise}},
  \code{\link{which.quantile}},
  \code{\link{whichEventsVincenty}},
  \code{\link{whichEvents}}, \code{\link{windrose}},
  \code{\link{workspace.clear}}, \code{\link{xkill}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
