% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.asp.r
\name{get.asp}
\alias{get.asp}
\title{get.asp}
\usage{
get.asp()
}
\description{
gets the aspect ratio
}
\seealso{
Other plotting: \code{\link{dist.plt}},
  \code{\link{dstamp}}, \code{\link{ellipse.2d}},
  \code{\link{legend.box}}, \code{\link{legend.bubble}},
  \code{\link{pie.add}}, \code{\link{pie.draw}},
  \code{\link{pie.legend}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
