% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_day.r
\name{get_day}
\alias{get_day}
\title{get_day}
\usage{
get_day(the.year, the.month)
}
\arguments{
\item{\code{the.year}}{YYYY}

\item{\code{the.month}}{MM}
}
\description{
This function provides the number of days in each month, given a year and month. This script was originally part of the date.picker() function, but was separated since it might be useful on its own
}
\note{
It accounts for leapyears from 1904-2096.  Leapyears are not simply every 4 years - please look it up should this script live long enough to need modification
}
\seealso{
Other date functions: \code{\link{chron2datestamp}},
  \code{\link{convert.datecodes}},
  \code{\link{date_picker}}, \code{\link{dayofyear}},
  \code{\link{get.day}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
