% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.box.r
\name{legend.box}
\alias{legend.box}
\title{legend.box}
\usage{
legend.box(x, y = NULL, maxradius, mab = 1.2, inset = 0, double = F)
}
\arguments{
\item{\code{x}}{position}

\item{\code{y}}{defaults to NULL}

\item{\code{maxradius}}{}

\item{\code{mab}}{defaults to 1.2}

\item{\code{inset}}{defaults to 0}

\item{\code{double}}{defaults to FALSE}
}
\description{
couldn't accurately describe
}
\seealso{
Other plotting: \code{\link{dist.plt}},
  \code{\link{dstamp}}, \code{\link{ellipse.2d}},
  \code{\link{get.asp}}, \code{\link{legend.bubble}},
  \code{\link{pie.add}}, \code{\link{pie.draw}},
  \code{\link{pie.legend}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
