% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.bubble.r
\name{legend.bubble}
\alias{legend.bubble}
\title{legend.bubble}
\usage{
legend.bubble(x, y = NULL, z, maxradius = 1, n = 3, round = 0,
  bty = "o", mab = 1.2, bg = NULL, inset = 0, pch = 21,
  pt.bg = NULL, txt.cex = 1, txt.col = NULL, font = NULL, ...)
}
\arguments{
\item{\code{x}}{}

\item{\code{y}}{defaults to NULL}

\item{\code{maxradius}}{defaults to 1}

\item{\code{n}}{defaults to 3}

\item{\code{round}}{defaults to 0}

\item{\code{bty}}{defaults to "o"}

\item{\code{mab}}{defaults to 1.2}

\item{\code{bg}}{defaults to NULL}

\item{\code{inset}}{defaults to 0}

\item{\code{pch}}{defaults to 21}

\item{\code{pt.bg}}{defaults to NULL}

\item{\code{txt.cex}}{defaults to 1}

\item{\code{txt.col}}{defaults to NULL}

\item{\code{font}}{defaults to NULL}
}
\description{
couldn't accurately describe
}
\seealso{
Other plotting: \code{\link{dist.plt}},
  \code{\link{dstamp}}, \code{\link{ellipse.2d}},
  \code{\link{get.asp}}, \code{\link{legend.box}},
  \code{\link{pie.add}}, \code{\link{pie.draw}},
  \code{\link{pie.legend}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
