% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.xyz.r
\name{make.xyz}
\alias{make.xyz}
\title{make.xyz}
\usage{
make.xyz(x, y, z, group, FUN = sum, ...)
}
\arguments{
\item{\code{x}}{}

\item{\code{y}}{}

\item{\code{z}}{}

\item{\code{group}}{}

\item{\code{FUN}}{}
}
\description{
couldn't accurately describe
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\keyword{manip}
