% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.sorted.angles.r
\name{nearest.sorted.angles}
\alias{nearest.sorted.angles}
\title{nearest.sorted.angles}
\usage{
nearest.sorted.angles(plist, hull, k, eps = pi/4, ub = NULL)
}
\description{
unknown
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applyMean}}, \code{\link{applySum}},
  \code{\link{break.list}}, \code{\link{by2}},
  \code{\link{colour.scale}}, \code{\link{compute.sums}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}, \code{\link{dim_list}},
  \code{\link{duplicates.toremove}}, \code{\link{na.zero}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
