% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.add.r
\name{pie.add}
\alias{pie.add}
\title{pie.add}
\usage{
pie.add(z, x = 0, y = 0, labels = names(z), radius = 1, edges = 200,
  clockwise = TRUE, init.angle = 90, density = NULL, angle = 45,
  col = NULL, border = NULL, lty = NULL, label.dist = 1.1, ...)
}
\arguments{
\item{\code{z}}{}

\item{\code{x}}{defaults to 0}

\item{\code{y}}{defaults to 0}

\item{\code{labels}}{defaults to names(z)}

\item{\code{radius}}{defaults to 1}

\item{\code{edges}}{defaults to 200}

\item{\code{clockwise}}{defaults to TRUE}

\item{\code{init.angle}}{defaults to 90}

\item{\code{density}}{defaults to NULL}

\item{\code{angle}}{defaults to 45}

\item{\code{col}}{defaults to NULL}

\item{\code{border}}{defaults to NULL}

\item{\code{lty}}{defaults to NULL}

\item{\code{label.dist}}{defaults to 1.1}
}
\description{
couldn't accurately describe
}
\seealso{
Other plotting: \code{\link{dist.plt}},
  \code{\link{dstamp}}, \code{\link{ellipse.2d}},
  \code{\link{get.asp}}, \code{\link{legend.box}},
  \code{\link{legend.bubble}}, \code{\link{pie.draw}},
  \code{\link{pie.legend}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
