% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.draw.r
\name{pie.draw}
\alias{pie.draw}
\title{pie.draw}
\usage{
pie.draw(x, y, z, radius, scale = T, labels = NA, silent = TRUE, ...)
}
\arguments{
\item{\code{x}}{NULL}

\item{\code{y}}{NULL}

\item{\code{z}}{NULL}

\item{\code{radius}}{NULL}

\item{\code{scale}}{NULL}

\item{\code{labels}}{NULL}

\item{\code{silent}}{NULL}
}
\description{
couldn't accurately describe
}
\note{
useage
data(landings)
data(coast)
xlim <- c(-12,-5)
ylim <- c(50,56)
xyz <- make.xyz(landings$Lon,landings$Lat,landings$LiveWeight,landings$Species)
col <- rainbow(5)
basemap(xlim, ylim, main = "Species composition of gadoid landings")
draw.shape(coast, col="cornsilk")
draw.pie(xyz$x, xyz$y, xyz$z, radius = 0.3, col=col)
legend.pie(-13.25,54.8,labels=c("cod","had","hke","pok","whg"), radius=0.3, bty="n", col=col,
 cex=0.8, label.dist=1.3)
legend.z <- round(max(rowSums(xyz$z,na.rm=TRUE))/10^6,0)
legend.bubble(-13.25,55.5,z=legend.z,round=1,maxradius=0.3,bty="n",txt.cex=0.6)
text(-12.25,56,"landings (kt)",cex=0.8)
}
\seealso{
Other plotting: \code{\link{dist.plt}},
  \code{\link{dstamp}}, \code{\link{ellipse.2d}},
  \code{\link{get.asp}}, \code{\link{legend.box}},
  \code{\link{legend.bubble}}, \code{\link{pie.add}},
  \code{\link{pie.legend}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
