% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.legend.r
\name{pie.legend}
\alias{pie.legend}
\title{pie.legend}
\usage{
pie.legend(x, y = NULL, z = NULL, labels, radius = 1, bty = "o",
  mab = 1.2, bg = NULL, inset = 0, ...)
}
\arguments{
\item{\code{x}}{NULL}

\item{\code{y}}{defaults to NULL}

\item{\code{z}}{defaults to NULL}

\item{\code{labels}}{NULL}

\item{\code{radius}}{defaults to 1}

\item{\code{bty}}{defaults to "o"}

\item{\code{mab}}{defaults to 1.2}

\item{\code{bg}}{defaults to NULL}

\item{\code{inset}}{defaults to 0}
}
\description{
couldnt accurately describe
}
\seealso{
Other plotting: \code{\link{dist.plt}},
  \code{\link{dstamp}}, \code{\link{ellipse.2d}},
  \code{\link{get.asp}}, \code{\link{legend.box}},
  \code{\link{legend.bubble}}, \code{\link{pie.add}},
  \code{\link{pie.draw}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
