% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatpoly_to_spatpolydf.r
\name{spatpoly_to_spatpolydf}
\alias{spatpoly_to_spatpolydf}
\title{MMM March 2016
This function takes a spatialpolygon object and converts it to a 
spatialpolygondataframe (which allows it to be exported to a shapefile via 
writeOGR()}
\usage{
spatpoly_to_spatpolydf(spat_poly)
}
\description{
MMM March 2016
This function takes a spatialpolygon object and converts it to a 
spatialpolygondataframe (which allows it to be exported to a shapefile via 
writeOGR()
}

