% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.r
\name{A}
\alias{A}
\title{A}
\usage{
A(x)
}
\description{
This function is a shorthand for \code{attr()}.
}
\seealso{
\code{\link[base]{attr}}

Other poorly documented: \code{\link{applyMean}},
  \code{\link{applySum}}, \code{\link{break.list}},
  \code{\link{by2}}, \code{\link{colour.scale}},
  \code{\link{compute.sums}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}, \code{\link{dim_list}},
  \code{\link{duplicates.toremove}}, \code{\link{na.zero}},
  \code{\link{nearest.sorted.angles}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
