% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyMean.r
\name{applyMean}
\alias{applyMean}
\title{applyMean}
\usage{
applyMean(f, method = "fast", newnames = NULL, ...)
}
\description{
apply a mean or weighted mean to data by factor
and return a formatted data.frame
must contain index, variable, weight in this sequence 
can take additional parmaters ...
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applySum}}, \code{\link{break.list}},
  \code{\link{by2}}, \code{\link{colour.scale}},
  \code{\link{compute.sums}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}, \code{\link{dim_list}},
  \code{\link{duplicates.toremove}}, \code{\link{na.zero}},
  \code{\link{nearest.sorted.angles}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
