% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applySum.r
\name{applySum}
\alias{applySum}
\title{applySum}
\usage{
applySum(f, method = "fast", newnames = NULL)
}
\description{
apply a sum and return a formated data frame. If 3 columns of data then a weighted sum is returned, if 1 column (vector) then a count is returned. Must contain index, variable, weight in this sequence
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other poorly documented, unclaimed: \code{\link{applyMean}}
}

