% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignStation.r
\name{assignStation}
\alias{assignStation}
\title{assignStation}
\usage{
assignStation(events, maxdist = 0.01, res = 0.005, expwin = 0.05,
  map = NULL, lines = F, ...)
}
\description{
couldn't accurately describe
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applySum}}, \code{\link{colour.scale}},
  \code{\link{compileODF}},
  \code{\link{compute.oxygen.saturation}},
  \code{\link{compute.sums}}, \code{\link{concave.hull}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}, \code{\link{cor.prob}},
  \code{\link{correlationMatrix}}
}

