% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break.list.r
\name{break.list}
\alias{break.list}
\title{@title break.list
 @description couldn't accurately describe
 @family poorly documented
 @author  unknown, \email{<unknown>@dfo-mpo.gc.ca}
 @export}
\usage{
break.list(X, delimit = NULL, return.columns = NULL, nvars = NULL)
}
\description{
@title break.list
 @description couldn't accurately describe
 @family poorly documented
 @author  unknown, \email{<unknown>@dfo-mpo.gc.ca}
 @export
}

