% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by2.r
\name{by2}
\alias{by2}
\title{by2}
\usage{
by2(x, indices, func, var, newvars, ...)
}
\description{
"slower but more flexible"
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applyMean}}, \code{\link{applySum}},
  \code{\link{break.list}}, \code{\link{colour.scale}},
  \code{\link{compute.sums}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}, \code{\link{dim_list}},
  \code{\link{duplicates.toremove}}, \code{\link{na.zero}},
  \code{\link{nearest.sorted.angles}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
