% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileODF.r
\name{compileODF}
\alias{compileODF}
\title{compileODF}
\usage{
compileODF(path = loc)
}
\arguments{
\item{\code{path}}{= location of an ODF file}
}
\value{
\code{result} =  what the function returns
}
\description{
Compile ODF files from in a specific folder (path) CTD casts into a data frame.  This script only requires casts to be labeled either down ('D') or up ('U') along the file naming convention
}
\note{
requires the \code{parse.odf.file()} function from this package.  Currently don't have an *.odf file to test this on
}
\author{
Adam Cook, \email{Adam.Cook@dfo-mpo.gc.ca}
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{assignStation}}, \code{\link{colour.scale}},
  \code{\link{compute.oxygen.saturation}},
  \code{\link{compute.sums}}, \code{\link{concave.hull}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}, \code{\link{cor.prob}},
  \code{\link{correlationMatrix}}
}

