% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeFun.r
\name{completeFun}
\alias{completeFun}
\title{completeFun}
\usage{
completeFun(data, desiredCols)
}
\arguments{
\item{\code{data}}{= the matrix/data.frame to be evaluated}

\item{\code{desiredCols}}{= identifier for the column that must have values}
}
\value{
\code{result} =  matrix/data.frame without rows that had \code{NA} in the \code{desiredCols}
}
\description{
This function removes rows where specific columns have NA, and is a function to return complete rows.
}
\note{
The name is deceptive - MMM did not find it to be fun at all.
}
\examples{
A = matrix(c(2, NA, 3, 1, 5, 7), nrow=2, ncol=3, byrow = TRUE)
A
[,1] [,2] [,3]
[1,]    2   NA    3
[2,]    1    5    7
completeFun(A, 2)
[1] 1 5 7
}
\author{
Adam Cook, \email{Adam.Cook@dfo-mpo.gc.ca}
}
