% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.sums.r
\name{compute.sums}
\alias{compute.sums}
\title{compute.sums}
\usage{
compute.sums(x, var, index)
}
\arguments{
\item{\code{x}}{}

\item{\code{var}}{}

\item{\code{index}}{}
}
\value{
\code{res}
}
\description{
couldn't accurately describe
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applySum}}, \code{\link{assignStation}},
  \code{\link{colour.scale}}, \code{\link{compileODF}},
  \code{\link{compute.oxygen.saturation}},
  \code{\link{concave.hull}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}, \code{\link{cor.prob}},
  \code{\link{correlationMatrix}}
}

