% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.sums.r
\name{compute.sums}
\alias{compute.sums}
\title{compute.sums}
\usage{
compute.sums(x, var, index)
}
\arguments{
\item{\code{x}}{=the input matrix}

\item{\code{var}}{=}

\item{\code{index}}{}
}
\value{
\code{res}
}
\description{
This function takes a matrix, and for a given column and index, returns a matrix of different unique values for each.
}
\note{
A = matrix(c(2, 4, 3, 1, 5, 7,2,4,7,2,5,6,1,2,3,5,6,7), ncol=3, byrow = TRUE)
compute.sums(A,"a",1) = for the unique values of first column, var1= corresponding value in col "a", sums of corresponding values in column a
compute.sums(A,"a",2) = for the unique values of second column, var1= corresponding value in col "a", Freq = sums of corresponding values in column a
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applyMean}}, \code{\link{applySum}},
  \code{\link{break.list}}, \code{\link{by2}},
  \code{\link{colour.scale}},
  \code{\link{convert.graphics.format}},
  \code{\link{convert2factor}}, \code{\link{dim_list}},
  \code{\link{duplicates.toremove}}, \code{\link{na.zero}},
  \code{\link{nearest.sorted.angles}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
