% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concave.hull.r
\name{concave.hull}
\alias{concave.hull}
\title{concave.hull}
\usage{
concave.hull(xy, k = 5, ub = NULL, dname = NULL, fname = NULL,
  random.start = F)
}
\arguments{
\item{\code{xy}}{}

\item{\code{k}}{}

\item{\code{ub}}{}

\item{\code{dname}}{}

\item{\code{fname}}{}
}
\description{
cannot accurately describe
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other abysmally documented: \code{\link{cor.prob}},
  \code{\link{correlationMatrix}},
  \code{\link{create.histograms.yearly}}
}

