% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concave.hull.r
\name{concave.hull}
\alias{concave.hull}
\title{concave.hull}
\usage{
concave.hull(xy, k = 5, ub = NULL, plot = FALSE, random.start = FALSE)
}
\arguments{
\item{\code{xy}}{Points data in x,y format}

\item{\code{k}}{Number of nearest neighbours to consider, defaults to 5}

\item{\code{ub}}{Upper bound in spatial distance to consider, defaults to 1/8 max range}

\item{\code{plot}}{defaults to TRUE}

\item{\code{random.start}}{defaults to FALSE otherwise the min values are used.}
}
\description{
Find the outline (boundary) of a set of points in 2D space.
}
\author{
Jae Choi, \email{<jae.choi>@dfo-mpo.gc.ca}
}
