% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.datecodes.r
\name{convert.datecodes}
\alias{convert.datecodes}
\title{convert.datecodes}
\usage{
convert.datecodes(x, y)
}
\arguments{
\item{\code{x}}{= a chron date object}

\item{\code{y}}{= the value to be extracted x, acceptable values include 'chron', 'year','month','day','week','julian','julian.ref.2000','tripcode'}
}
\value{
\code{z} =  the requested date part
}
\description{
This function take a chron object x and extract required data
}
\examples{
t=Sys.Date()
convert.datecodes(t,'year')
[1] 2016
convert.datecodes(t,'week')
[1] 25
}
\seealso{
Other date functions: \code{\link{chron2datestamp}},
  \code{\link{date_picker}}, \code{\link{dayofyear}},
  \code{\link{get.day}}, \code{\link{get_day}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
