% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.graphics.format.r
\name{convert.graphics.format}
\alias{convert.graphics.format}
\title{convert.graphics.format}
\usage{
convert.graphics.format(file0, file1, options = "")
}
\arguments{
\item{\code{file0}}{}

\item{\code{file1}}{}

\item{\code{options}}{}
}
\value{
\code{}
}
\description{
this uses imagemagicks functionality
}
\note{
does not work in Windows
}
\references{
\url{http://www.imagemagick.org/script/convert.php}
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applyMean}}, \code{\link{applySum}},
  \code{\link{break.list}}, \code{\link{by2}},
  \code{\link{colour.scale}}, \code{\link{compute.sums}},
  \code{\link{convert2factor}}, \code{\link{dim_list}},
  \code{\link{duplicates.toremove}}, \code{\link{na.zero}},
  \code{\link{nearest.sorted.angles}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
