% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2factor.r
\name{convert2factor}
\alias{convert2factor}
\title{convert2factor}
\usage{
convert2factor(x, vars)
}
\arguments{
\item{\code{x}}{}

\item{\code{vars}}{}
}
\value{
\code{}
}
\description{
converts objects to factors
}
\note{
does not work in Windows
}
\seealso{
Other poorly documented: \code{\link{A}},
  \code{\link{applyMean}}, \code{\link{applySum}},
  \code{\link{break.list}}, \code{\link{by2}},
  \code{\link{colour.scale}}, \code{\link{compute.sums}},
  \code{\link{convert.graphics.format}},
  \code{\link{dim_list}},
  \code{\link{duplicates.toremove}}, \code{\link{na.zero}},
  \code{\link{nearest.sorted.angles}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
