% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.prob.r
\name{cor.prob}
\alias{cor.prob}
\title{cor.prob}
\usage{
cor.prob(X, dfr = nrow(X) - 2)
}
\arguments{
\item{\code{X}}{}

\item{\code{dfr}}{degrees of freedom}
}
\value{
\code{}
}
\description{
couldn't accurately describe
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
\seealso{
Other abysmally documented: \code{\link{concave.hull}},
  \code{\link{correlationMatrix}},
  \code{\link{create.histograms.yearly}}
}

