% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMatrix.r
\name{correlationMatrix}
\alias{correlationMatrix}
\title{correlationMatrix}
\usage{
correlationMatrix(X, dfr = nrow(X) - 2, meth. = c("pearson", "kendall",
  "spearman"))
}
\arguments{
\item{\code{X}}{}

\item{\code{dfr}}{degrees of freedom}

\item{\code{meth.}}{}
}
\value{
\code{}
}
\description{
couldn't accurately describe
}
\note{
X is the matrix of values with each column representing a single variable.  The upper triangle of the resulting matrix is p-values and lower is cor test value
}
\seealso{
Other statistical: \code{\link{devianceGLM}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
