% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataOverlap.r
\name{dataOverlap}
\alias{dataOverlap}
\title{dataOverlap}
\usage{
dataOverlap(data1, data2, distance.Theshold = 0.1, dnames = c("EID", "X",
  "Y", "date", "salinity", "temperature", "depth", "source"))
}
\arguments{
\item{\code{data1}}{dataset #1}

\item{\code{data2}}{dataset #2}

\item{\code{distance.Theshold}}{defaults to 0.1}

\item{\code{dnames}}{defaults to \code{c("EID","X","Y","date","salinity","temperature","depth","source")}}
}
\value{
\code{Data}
}
\description{
compare 2 datasets and returns values that overlap
}
\note{
should not the return object should probably be Data1, not Data?
}
\seealso{
Other overly specific: \code{\link{cutOutLand}},
  \code{\link{default.lattice.options}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
