% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayofyear.r
\name{dayofyear}
\alias{dayofyear}
\title{dayofyear}
\usage{
dayofyear(a)
}
\arguments{
\item{\code{t}}{= the date}
}
\value{
= the numerical day of the year
}
\description{
This function converts dates from \code{chron::} format into the day number of year
}
\examples{
dayofyear("2016-02-11")
[1] 42
}
\seealso{
Other date functions: \code{\link{chron2datestamp}},
  \code{\link{convert.datecodes}},
  \code{\link{date_picker}}, \code{\link{get.day}},
  \code{\link{get_day}}
}
\author{
unknown, \email{<unknown>@dfo-mpo.gc.ca}
}
